// LDC.Atlas.WebApp/Client/app/shared/entities/commodity.entity.ts
export interface Commodity {
    commodityId: number;
    principalCommodity: string;
    commodityType: string;
    part2: string;
    part3: string;
    part4: string;
    part5: string;
    description: string;
    arbitrationCode: string;
    currency: string;
    weightCode: string;
    priceCode: string;
    weightUnitId: number;
    priceUnitId: number;
    mDMId: number;
    countryCode: string;
}